requirements.txt
# Core ML & Data
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization (for analysis)
matplotlib>=3.7.0
seaborn>=0.12.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# MLOps & Model Management
mlflow>=2.8.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Monitoring & Logging
python-json-logger>=2.0.0

# Utilities
python-dotenv>=1.0.0

# UI Framework
streamlit>=1.28.0
requests>=2.31.0

# Testing
pytest>=7.4.0
httpx>=0.24.0

# Code Quality
black>=23.0.0
isort>=5.12.0
ruff>=0.1.0
